fix: removed unwanted code and merged user and organization components
