+++ web/src/components/iam/organizations/AppOrganizations.vue
@@ -24,7 +24,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 import { defineComponent, ref } from "vue";
 import { useStore } from "vuex";
 import { useI18n } from "vue-i18n";
-import OrganizationsCloud from "@/enterprise/components/organizations/Organization.vue";
 import OrganizationsEnterprise from "@/components/iam/organizations/ListOrganizations.vue";
 
 import config from "@/aws-exports";
@@ -33,7 +32,6 @@ import { watch } from "vue";
 export default defineComponent({
   name: "AppOrganizations",
   components: {
-    OrganizationsCloud,
     OrganizationsEnterprise,
   },
   setup() {
@@ -47,11 +45,11 @@ export default defineComponent({
     watch(
       () => store.state.zoConfig,
       (zoConfig) => {
-        if (config.isCloud == "true") {
-          componentName.value = "OrganizationsCloud";
-        }
-
-        if (zoConfig.sso_enabled || config.isEnterprise == "true") {
+        if (
+          zoConfig.sso_enabled ||
+          config.isEnterprise == "true" ||
+          config.isCloud == "true"
+        ) {
           componentName.value = "OrganizationsEnterprise";
         }
 
@@ -59,7 +57,7 @@ export default defineComponent({
       },
       {
         immediate: true,
-      }
+      },
     );
 
     return { store, t, componentName, loadComponent };
