+++ web/src/components/iam/serviceAccounts/AddServiceAccount.vue
@@ -122,7 +122,6 @@ import {
   getImageURL,
   useLocalCurrentUser,
   useLocalUserInfo,
-  getLogoutURL,
   invlidateLoginData,
 } from "@/utils/zincutils";
 import config from "@/aws-exports";
