+++ web/src/components/iam/users/User.vue
@@ -96,8 +96,13 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
               <q-icon name="search" />
             </template>
           </q-input>
-
-          <div class="col-6">
+          <div class="col-6" v-if="config.isCloud == 'true'">
+            <member-invitation
+              :key="currentUserRole"
+              v-model:currentrole="currentUserRole"
+            />
+          </div>
+          <div class="col-6" v-else>
             <q-btn
               v-if="showAddUserBtn"
               class="q-ml-md q-mb-xs text-bold no-border"
@@ -166,7 +171,9 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
         </q-card-section>
 
         <q-card-actions class="confirmActions">
-          <q-btn v-close-popup="true" unelevated no-caps class="q-mr-sm">
+          <q-btn v-close-popup="true"
+unelevated no-caps
+class="q-mr-sm">
             {{ t("user.cancel") }}
           </q-btn>
           <q-btn
@@ -199,6 +206,7 @@ import AddUser from "@/components/iam/users/AddUser.vue";
 import NoData from "@/components/shared/grid/NoData.vue";
 import organizationsService from "@/services/organizations";
 import segment from "@/services/segment_analytics";
+import MemberInvitation from "@/components/iam/users/MemberInvitation.vue";
 import {
   getImageURL,
   verifyOrganizationStatus,
@@ -213,7 +221,13 @@ import { getRoles } from "@/services/iam";
 
 export default defineComponent({
   name: "UserPageOpenSource",
-  components: { QTablePagination, UpdateUserRole, NoData, AddUser },
+  components: {
+    QTablePagination,
+    UpdateUserRole,
+    NoData,
+    AddUser,
+    MemberInvitation,
+  },
   emits: [
     "updated:fields",
     "deleted:fields",
@@ -331,14 +345,7 @@ export default defineComponent({
 
       return new Promise((resolve, reject) => {
         usersService
-          .orgUsers(
-            0,
-            1000,
-            "email",
-            false,
-            "",
-            store.state.selectedOrganization.identifier
-          )
+          .orgUsers(store.state.selectedOrganization.identifier)
           .then((res) => {
             resultTotal.value = res.data.data.length;
             let counter = 1;
@@ -361,11 +368,11 @@ export default defineComponent({
                 role: data.role,
                 member_created: date.formatDate(
                   parseInt(data.member_created),
-                  "YYYY-MM-DDTHH:mm:ssZ"
+                  "YYYY-MM-DDTHH:mm:ssZ",
                 ),
                 member_updated: date.formatDate(
                   parseInt(data.member_updated),
-                  "YYYY-MM-DDTHH:mm:ssZ"
+                  "YYYY-MM-DDTHH:mm:ssZ",
                 ),
                 org_member_id: data.org_member_id,
                 isLoggedinUser: store.state.userInfo.email == data.email,
@@ -533,7 +540,7 @@ export default defineComponent({
           {
             row: props.row,
           },
-          true
+          true,
         );
       } else {
         addUser({}, false);
@@ -639,7 +646,7 @@ export default defineComponent({
           }
         })
         .catch((err: any) => {
-          if(err.response.status != 403){
+          if (err.response.status != 403) {
             $q.notify({
               color: "negative",
               message: "Error while deleting user.",
@@ -663,7 +670,7 @@ export default defineComponent({
             email: row.email,
             organization_id: parseInt(store.state.selectedOrganization.id),
           },
-          store.state.selectedOrganization.identifier
+          store.state.selectedOrganization.identifier,
         )
         .then((res: { data: any }) => {
           if (res.data.error_members != null) {
